backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
sqlalchemy==2.0.30
psycopg[binary]==3.1.19
python-multipart==0.0.9
redis==5.0.4
rq==1.16.2
boto3==1.34.131
python-jose==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
stripe==10.8.0
edge-tts==7.2.3

pytest==8.2.2
